gab created test

+++ walkTheDog/test/byui/cit260/walkTheDog/control/EventControlTest.java
@@ -128,7 +128,7 @@ public class EventControlTest {
         System.out.println("\tTest case #1");
         double noExploreCounter = 5.0;
         int idealLeashLength = 7;
-        double userDefinedNumber = 7.0;
+        int userDefinedNumber = 7;
         EventControl instance = new EventControl();
         int expResult = 1;
         int result = instance.eventOnNoExplore(noExploreCounter, idealLeashLength, userDefinedNumber);
@@ -140,7 +140,7 @@ public class EventControlTest {
         System.out.println("\tTest case #2");
         noExploreCounter = 1.0;
         idealLeashLength = -6;
-        userDefinedNumber = 5.0;
+        userDefinedNumber = 5;
         expResult = -1;
         result = instance.eventOnNoExplore(noExploreCounter, idealLeashLength, userDefinedNumber);
         assertEquals(expResult, result);
@@ -151,7 +151,7 @@ public class EventControlTest {
         System.out.println("\tTest case #3");
         noExploreCounter = 2.0;
         idealLeashLength = 7;
-        userDefinedNumber = 19.0;
+        userDefinedNumber = 19;
         expResult = 1;
         result = instance.eventOnNoExplore(noExploreCounter, idealLeashLength, userDefinedNumber);
         assertEquals(expResult, result);
@@ -162,7 +162,7 @@ public class EventControlTest {
         System.out.println("\tTest case #4");
         noExploreCounter = 0.0;
         idealLeashLength = 4;
-        userDefinedNumber = 19.0;
+        userDefinedNumber = 19;
         expResult = 1;
         result = instance.eventOnNoExplore(noExploreCounter, idealLeashLength, userDefinedNumber);
         assertEquals(expResult, result);
@@ -173,7 +173,7 @@ public class EventControlTest {
         System.out.println("\tTest case #5");
         noExploreCounter = 10.0;
         idealLeashLength = 15;
-        userDefinedNumber = 19.0;
+        userDefinedNumber = 19;
         expResult = 1;
         result = instance.eventOnNoExplore(noExploreCounter, idealLeashLength, userDefinedNumber);
         assertEquals(expResult, result);
@@ -184,7 +184,7 @@ public class EventControlTest {
         System.out.println("\tTest case #6");
         noExploreCounter = 1.0;
         idealLeashLength = 4;
-        userDefinedNumber = 15.0;
+        userDefinedNumber = 15;
         expResult = 1;
         result = instance.eventOnNoExplore(noExploreCounter, idealLeashLength, userDefinedNumber);
         assertEquals(expResult, result);
